implemented product support, needs more work

+++ src/main/java/ba/unsa/etf/si/mainserver/controllers/ProductController.java
@@ -6,13 +6,16 @@ import ba.unsa.etf.si.mainserver.models.business.Office;
 import ba.unsa.etf.si.mainserver.models.products.Discount;
 import ba.unsa.etf.si.mainserver.models.products.OfficeInventory;
 import ba.unsa.etf.si.mainserver.models.products.Product;
+import ba.unsa.etf.si.mainserver.requests.products.DiscountRequest;
 import ba.unsa.etf.si.mainserver.requests.products.InventoryRequest;
 import ba.unsa.etf.si.mainserver.requests.products.ProductRequest;
 import ba.unsa.etf.si.mainserver.responses.ApiResponse;
+import ba.unsa.etf.si.mainserver.responses.products.DiscountResponse;
 import ba.unsa.etf.si.mainserver.responses.products.OfficeInventoryResponse;
 import ba.unsa.etf.si.mainserver.responses.products.ProductResponse;
 import ba.unsa.etf.si.mainserver.services.business.BusinessService;
 import ba.unsa.etf.si.mainserver.services.business.OfficeService;
+import ba.unsa.etf.si.mainserver.services.products.DiscountService;
 import ba.unsa.etf.si.mainserver.services.products.OfficeInventoryService;
 import ba.unsa.etf.si.mainserver.services.products.ProductService;
 import org.springframework.http.ResponseEntity;
@@ -31,12 +34,14 @@ public class ProductController {
     private final BusinessService businessService;
     private final OfficeService officeService;
     private final OfficeInventoryService officeInventoryService;
+    private final DiscountService discountService;
 
-    public ProductController(ProductService productService, BusinessService businessService, OfficeService officeService, OfficeInventoryService officeInventoryService) {
+    public ProductController(ProductService productService, BusinessService businessService, OfficeService officeService, OfficeInventoryService officeInventoryService, DiscountService discountService) {
         this.productService = productService;
         this.businessService = businessService;
         this.officeService = officeService;
         this.officeInventoryService = officeInventoryService;
+        this.discountService = discountService;
     }
 
     @GetMapping("/products")
@@ -44,7 +49,6 @@ public class ProductController {
     public List<ProductResponse> getAllProductsForBusiness(@PathVariable("businessId") Long businessId){
         Optional<Business> businessOptional = businessService.findById(businessId);
         if(businessOptional.isPresent()){
-            //System.out.println(businessOptional.get().getName());
             return businessOptional.get().getProducts().stream().map(ProductResponse::new).collect(Collectors.toList());
         }
 
@@ -81,7 +85,7 @@ public class ProductController {
                         productRequest.getUnit(),
                         productRequest.getImage());
 
-                Discount discount = new Discount(product, 0);
+                Discount discount = new Discount(0);
                 product.setBusiness(businessOptional.get());
                 product.setDiscount(discount);
                 businessService.save(businessOptional.get());
@@ -119,14 +123,17 @@ public class ProductController {
                                                    @PathVariable("productId") Long productId){
         Optional<Business> businessOptional = businessService.findById(businessId);
         if(businessOptional.isPresent()){
-            Optional<Product> productOptional = productService.findByBusiness(businessOptional.get());
+            Optional<Product> productOptional = productService.findById(productId);
             if(productOptional.isPresent()){
-                productService.delete(productOptional.get());
-                return ResponseEntity.ok(new ApiResponse("Product successfully deleted", 200));
+                if(productOptional.get().getBusiness().getId().equals(businessId)){
+                    officeInventoryService.findByProduct(productOptional.get()).forEach(officeInventoryService::delete);
+                    productService.delete(productOptional.get());
+                    return ResponseEntity.ok(new ApiResponse("Product successfully deleted", 200));
+                }
+                throw new AppException("Product with id " + productId + " doesn't exist for business with id" + businessId);
             }
-            throw new AppException("Product with id " + productId + " doesn't exist for business with id" + businessId);
+            throw new AppException("Product with id " + productId + " doesn't exist");
         }
-
         throw new AppException("Business with id " + businessId + " doesn't exist");
     }
 
@@ -137,19 +144,21 @@ public class ProductController {
 
         Optional<Business> businessOptional = businessService.findById(businessId);
         if(businessOptional.isPresent()){
-            Optional<Product> productOptional = productService.findByBusiness(businessOptional.get());
+            Optional<Product> productOptional = productService.findById(inventoryRequest.getProductId());
             if(productOptional.isPresent()){
-                Optional<Office> officeOptional = officeService.findByBusiness(businessOptional.get());
+                if(productOptional.get().getBusiness().getId().equals(businessId)){
+                    Optional<Office> officeOptional = officeService.findByBusiness(businessOptional.get());
                 if(officeOptional.isPresent()){
                     OfficeInventory officeInventory = new OfficeInventory(officeOptional.get(),
                             productOptional.get(),
                             inventoryRequest.getQuantity());
-
                     return new OfficeInventoryResponse(officeInventoryService.save(officeInventory));
                 }
                 throw new AppException("Office with id " + inventoryRequest.getOfficeId() + " doesn't exist for business with id" + businessId);
+                }
+                throw new AppException("Product with id " + inventoryRequest.getProductId() + " doesn't exist for business with id" + businessId);
             }
-            throw new AppException("Product with id " + inventoryRequest.getProductId() + " doesn't exist for business with id" + businessId);
+            throw new AppException("Product with id " + inventoryRequest.getProductId() + " doesn't exist");
         }
 
         throw new AppException("Business with id " + businessId + " doesn't exist");
@@ -158,30 +167,56 @@ public class ProductController {
     @PutMapping("/inventory")
     @Secured("ROLE_ADMIN")
     public OfficeInventoryResponse updateInventoryForBusiness(@PathVariable("businessId") Long businessId,
-                                                               @RequestBody InventoryRequest inventoryRequest){
+                                                               @RequestBody InventoryRequest inventoryRequest) {
+        Optional<Business> businessOptional = businessService.findById(businessId);
+        if (businessOptional.isPresent()) {
+            Optional<Product> productOptional = productService.findById(inventoryRequest.getProductId());
+            if (productOptional.isPresent()) {
+                if (productOptional.get().getBusiness().getId().equals(businessId)) {
+                    Optional<Office> officeOptional = officeService.findByBusiness(businessOptional.get());
+                    if (officeOptional.isPresent()) {
+                        Optional<OfficeInventory> officeInventoryOptional = officeInventoryService.
+                                findByProductAndOffice(productOptional.get(), officeOptional.get());
+                        if (officeInventoryOptional.isPresent()) {
+                            officeInventoryOptional.get().setOffice(officeOptional.get());
+                            officeInventoryOptional.get().setProduct(productOptional.get());
+                            officeInventoryOptional.get().setQuantity(inventoryRequest.getQuantity());
+                            return new OfficeInventoryResponse(officeInventoryService.save(officeInventoryOptional.get()));
+                        }
+                        throw new AppException("Office with id " + inventoryRequest.getOfficeId() + " doesn't exist for business with id" + businessId);
+                    }
+                    throw new AppException("Product with id " + inventoryRequest.getProductId() + " doesn't exist for business with id" + businessId);
+                }
+                throw new AppException("Product with id " + inventoryRequest.getProductId() + " doesn't exist");
+            }
+        }
+        throw new AppException("Business with id " + businessId + " doesn't exist");
+    }
+
+//ova ruta ne valja zato sto uvijek dodam discount kad dodajem proizvod pa ga treba samo promijeniti
+    //ne valja i kad uradim get nakon ovoga
+    @PutMapping("/products/{productId}/discount")
+    @Secured("ROLE_ADMIN")
+    public DiscountResponse addDiscount(@PathVariable("businessId") Long businessId,
+                                        @PathVariable("productId") Long productId,
+                                        @RequestBody DiscountRequest discountRequest){
         Optional<Business> businessOptional = businessService.findById(businessId);
         if(businessOptional.isPresent()){
-            Optional<Product> productOptional = productService.findByBusiness(businessOptional.get());
-            if(productOptional.isPresent()){
-                Optional<Office> officeOptional = officeService.findByBusiness(businessOptional.get());
-                if(officeOptional.isPresent()){
-                    Optional<OfficeInventory> officeInventoryOptional = officeInventoryService.
-                            findByProductAndOffice(productOptional.get(),officeOptional.get());
-                    if(officeInventoryOptional.isPresent()){
-                        officeInventoryOptional.get().setOffice(officeOptional.get());
-                        officeInventoryOptional.get().setProduct(productOptional.get());
-                        officeInventoryOptional.get().setQuantity(inventoryRequest.getQuantity());
-                        return new OfficeInventoryResponse(officeInventoryService.save(officeInventoryOptional.get()));
+            Optional<Product> productOptional = productService.findById(productId);
+            if (productOptional.isPresent()){
+                if(productOptional.get().getBusiness().getId().equals(businessId)){
+                    Optional<Discount> discountOptional = discountService.findById(productOptional.get().getDiscount().getId());
+                    if(discountOptional.isPresent()){
+                        System.out.println(discountOptional.get().getId());
+                        discountOptional.get().setPercentage(discountRequest.getPercentage());
+                        return new DiscountResponse(discountService.save(discountOptional.get()));
                     }
-                    throw new AppException("Office with id " + inventoryRequest.getOfficeId() + " doesn't exist for business with id" + businessId);
                 }
-                throw new AppException("Office with id " + inventoryRequest.getOfficeId() + " doesn't exist for business with id" + businessId);
+                throw new AppException("Product with id " + productId+ " doesn't exist for business with id" + businessId);
             }
-            throw new AppException("Product with id " + inventoryRequest.getProductId() + " doesn't exist for business with id" + businessId);
+            throw new AppException("Product with id " + productId+ " doesn't exist");
         }
 
         throw new AppException("Business with id " + businessId + " doesn't exist");
     }
-
-
 }
